[Refactor] Move AgentStatus to common header
