+++ be/src/storage/task/engine_batch_load_task.h
@@ -37,7 +37,7 @@
 #include <utility>
 #include <vector>
 
-#include "agent/status.h"
+#include "common/agent_status.h"
 #include "gen_cpp/AgentService_types.h"
 #include "gen_cpp/MasterService_types.h"
 #include "storage/olap_common.h"
